MOEN-45983: Removed old architecture support
