+++ SampleApp/android/gradle.properties
@@ -36,12 +36,5 @@ org.gradle.jvmargs=-Xmx4096M
 # ./gradlew <task> -PreactNativeArchitectures=x86_64
 reactNativeArchitectures=armeabi-v7a,arm64-v8a,x86,x86_64
 
-# Use this property to enable support to the new architecture.
-# This will allow you to use TurboModules and the Fabric render in
-# your application. You should enable this flag either if you want
-# to write custom TurboModules/Fabric components OR use libraries that
-# are providing them.
-newArchEnabled=true
-
 android.builtInKotlin=true
 android.newDsl=true
