+++ SampleApp/package.json
@@ -20,11 +20,11 @@
     "react-native": "0.87.0",
     "react-native-alertbox": "^0.2.11",
     "react-native-elements": "^3.4.3",
-    "react-native-moengage": "file:../sdk/core",
-    "react-native-moengage-cards": "file:../sdk/cards",
-    "react-native-moengage-geofence": "file:../sdk/geofence",
-    "react-native-moengage-inbox": "file:../sdk/inbox",
-    "react-native-moengage-personalize": "file:../sdk/personalize",
+    "react-native-moengage": "^12.1.0",
+    "react-native-moengage-cards": "^6.0.0",
+    "react-native-moengage-geofence": "^7.0.0",
+    "react-native-moengage-inbox": "^6.0.0",
+    "react-native-moengage-personalize": "^1.0.0",
     "react-native-render-html": "^6.3.4",
     "react-native-safe-area-context": "^5.5.2",
     "react-native-screens": "4.27.0",
