+++ sdk/cards/ReactNativeMoEngageCards.podspec
@@ -23,10 +23,6 @@ Pod::Spec.new do |s|
   s.dependency 'ReactNativeMoEngage'
   s.module_map = false
 
-  if defined?(install_modules_dependencies()) != nil
-    install_modules_dependencies(s);
-  else
-    s.dependency "React-Core"
-  end
+  install_modules_dependencies(s)
 end
