--- sdk/cards/android/src/oldarch/java/com/moengage/react/cards/MoEngageCardsBridge.kt
@@ -1,126 +1,0 @@
-/*
- * Copyright (c) 2014-2024 MoEngage Inc.
- *
- * All rights reserved.
- *
- *  Use of source code or binaries contained within MoEngage SDK is permitted only to enable use of the MoEngage platform by customers of MoEngage.
- *  Modification of source code and inclusion in mobile apps is explicitly allowed provided that all other conditions are met.
- *  Neither the name of MoEngage nor the names of its contributors may be used to endorse or promote products derived from this software without specific prior written permission.
- *  Redistribution of source code or binaries is disallowed except with specific prior written permission. Any such redistribution must retain the above copyright notice, this list of conditions and the following disclaimer.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-package com.moengage.react.cards
-
-import android.content.Context
-import com.facebook.react.bridge.Promise
-import com.facebook.react.bridge.ReactApplicationContext
-import com.facebook.react.bridge.ReactContextBaseJavaModule
-import com.facebook.react.bridge.ReactMethod
-import com.moengage.platform.internal.logger.Logger
-import com.moengage.platform.internal.logger.PlatformLogLevel
-import com.moengage.plugin.base.cards.CardsPluginHelper
-import com.moengage.plugin.base.cards.internal.cardListToJson
-import com.moengage.plugin.base.cards.internal.setCardsEventEmitter
-import org.json.JSONObject
-
-/**
- * Bridge to communicate with React-Native Cards Plugin in old arch
- *
- * @author Abhishek Kumar
- * @since Todo: Add Version
- */
-class MoEngageCardsBridge(
-    private val reactContext: ReactApplicationContext
-) : ReactContextBaseJavaModule(reactContext) {
-
-    private val bridgeHandler = MoEngageCardsBridgeHandler(reactContext)
-
-    override fun getName() = bridgeHandler.getName()
-
-    @ReactMethod
-    fun addListener(eventName: String) {
-        // Keep: Required for RN built in Event Emitter Calls.
-    }
-
-    @ReactMethod
-    fun removeListeners(count: Int) {
-        // Keep: Required for RN built in Event Emitter Calls.
-    }
-
-    @ReactMethod
-    fun initialize(payload: String) {
-        bridgeHandler.initialize(payload)
-    }
-
-    @ReactMethod
-    fun refreshCards(payload: String) {
-        bridgeHandler.refreshCards(payload)
-    }
-
-    @ReactMethod
-    fun onCardSectionLoaded(payload: String) {
-        bridgeHandler.onCardSectionLoaded(payload)
-    }
-
-    @ReactMethod
-    fun onCardSectionUnLoaded(payload: String) {
-        bridgeHandler.onCardSectionUnLoaded(payload)
-    }
-
-    @ReactMethod
-    fun getCardsCategories(payload: String, promise: Promise) {
-        bridgeHandler.getCardsCategories(payload, promise)
-    }
-
-    @ReactMethod
-    fun getCardsInfo(payload: String, promise: Promise) {
-        bridgeHandler.getCardsInfo(payload, promise)
-    }
-
-    @ReactMethod
-    fun cardClicked(payload: String) {
-        bridgeHandler.cardClicked(payload)
-    }
-
-    @ReactMethod
-    fun cardDelivered(payload: String) {
-        bridgeHandler.cardDelivered(payload)
-    }
-
-    @ReactMethod
-    fun cardShown(payload: String) {
-        bridgeHandler.cardShown(payload)
-    }
-
-    @ReactMethod
-    fun getCardsForCategory(payload: String, promise: Promise) {
-        bridgeHandler.getCardsForCategory(payload, promise)
-    }
-
-    @ReactMethod
-    fun deleteCards(payload: String) {
-        bridgeHandler.deleteCards(payload)
-    }
-
-    @ReactMethod
-    fun isAllCategoryEnabled(payload: String, promise: Promise) {
-        bridgeHandler.isAllCategoryEnabled(payload, promise)
-    }
-
-    @ReactMethod
-    fun getNewCardsCount(payload: String, promise: Promise) {
-        bridgeHandler.getNewCardsCount(payload, promise)
-    }
-
-    @ReactMethod
-    fun getUnClickedCardsCount(payload: String, promise: Promise) {
-        bridgeHandler.getUnClickedCardsCount(payload, promise)
-    }
-
-    @ReactMethod
-    fun fetchCards(payload: String, promise: Promise) {
-        bridgeHandler.fetchCards(payload, promise)
-    }
-}
