+++ sdk/cards/ios/MoEngageCardsBridge.mm
@@ -93,10 +93,8 @@ RCT_EXPORT_METHOD(deleteCards:(NSString *) payload) {
     [[MoEReactNativeCardsHandler sharedInstance] deleteCards:payload];
 }
 
-#ifdef RCT_NEW_ARCH_ENABLED
 - (std::shared_ptr<facebook::react::TurboModule>)getTurboModule:(const facebook::react::ObjCTurboModule::InitParams &)params {
     return std::make_shared<facebook::react::NativeMoEngageCardsSpecJSI>(params);
 }
-#endif
 
 @end
