+++ sdk/core/ReactNativeMoEngage.podspec
@@ -35,9 +35,5 @@ Pod::Spec.new do |s|
   CMD
 
 
-  if defined?(install_modules_dependencies()) != nil
-    install_modules_dependencies(s);
-  else
-    s.dependency "React-Core"
-  end
+  install_modules_dependencies(s)
 end
