+++ sdk/core/iOS/MoEReactBridge/MoEReactBridge.h
@@ -10,12 +10,7 @@
 #import <React/RCTBridgeModule.h>
 #import <React/RCTEventEmitter.h>
 #import "MoEReactEventDispatcher.h"
-
-#ifdef RCT_NEW_ARCH_ENABLED
 #import <NativeMoEngageSpec/NativeMoEngageSpec.h>
+
 @interface MoEReactBridge : RCTEventEmitter <NativeMoEngageSpec, MoEReactEventDispatcher>
 @end
-#else
-@interface MoEReactBridge : RCTEventEmitter <RCTBridgeModule, MoEReactEventDispatcher>
-@end
-#endif
