+++ sdk/core/iOS/MoEReactBridge/MoEReactBridge.mm
@@ -259,10 +259,8 @@ RCT_EXPORT_METHOD(deviceIdentifierTrackingStatusUpdate:(NSString *)payload) {
 }
 
 
-#ifdef RCT_NEW_ARCH_ENABLED
 - (std::shared_ptr<facebook::react::TurboModule>)getTurboModule:(const facebook::react::ObjCTurboModule::InitParams &)params {
   return std::make_shared<facebook::react::NativeMoEngageSpecJSI>(params);
 }
-#endif
 
 @end
