+++ sdk/geofence/ReactNativeMoEngageGeofence.podspec
@@ -21,9 +21,5 @@ Pod::Spec.new do |s|
   s.module_map = false
   
 
-  if defined?(install_modules_dependencies()) != nil
-    install_modules_dependencies(s);
-  else
-    s.dependency "React-Core"
-  end
+  install_modules_dependencies(s)
 end
