+++ sdk/geofence/iOS/MoEReactGeofence.mm
@@ -19,10 +19,8 @@ RCT_EXPORT_METHOD(stopGeofenceMonitoring:(NSString *) payload) {
     [[MoEReactNativeGeofenceHandler sharedInstance] stopGeofenceMonitoring:payload];
 }
 
-#ifdef RCT_NEW_ARCH_ENABLED
 - (std::shared_ptr<facebook::react::TurboModule>)getTurboModule:(const facebook::react::ObjCTurboModule::InitParams &)params {
   return std::make_shared<facebook::react::NativeMoEngageGeofenceSpecJSI>(params);
 }
-#endif
 
 @end
