--- sdk/inbox/android/src/oldarch/java/com/moengage/react/inbox/MoEReactInbox.kt
@@ -1,60 +1,0 @@
-/*
- * Copyright (c) 2014-2024 MoEngage Inc.
- *
- * All rights reserved.
- *
- *  Use of source code or binaries contained within MoEngage SDK is permitted only to enable use of the MoEngage platform by customers of MoEngage.
- *  Modification of source code and inclusion in mobile apps is explicitly allowed provided that all other conditions are met.
- *  Neither the name of MoEngage nor the names of its contributors may be used to endorse or promote products derived from this software without specific prior written permission.
- *  Redistribution of source code or binaries is disallowed except with specific prior written permission. Any such redistribution must retain the above copyright notice, this list of conditions and the following disclaimer.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
- 
-package com.moengage.react.inbox
-
-import com.facebook.react.bridge.Promise
-import com.facebook.react.bridge.ReactApplicationContext
-import com.facebook.react.bridge.ReactContextBaseJavaModule
-import com.facebook.react.bridge.ReactMethod
-import com.moengage.platform.internal.logger.Logger
-import com.moengage.platform.internal.logger.PlatformLogLevel
-import com.moengage.plugin.base.inbox.internal.InboxPluginHelper
-import com.moengage.plugin.base.inbox.internal.inboxDataToJson
-import com.moengage.plugin.base.inbox.internal.unClickedCountToJson
-
-/**
- * Bridge to communicate with inbox plugin js code in old arch
- *
- * @author Abhishek Kumar
- * @since Todo: Add Version
- */
-class MoEReactInbox(
-    reactContext: ReactApplicationContext
-) : ReactContextBaseJavaModule(reactContext) {
-
-    private val bridgeHandler = MoEngageInboxHandler(reactContext.applicationContext)
-
-
-    override fun getName() = bridgeHandler.getName()
-
-    @ReactMethod
-    fun getUnClickedCount(payload: String, promise: Promise) {
-        bridgeHandler.getUnClickedCount(payload, promise)
-    }
-
-    @ReactMethod
-    fun fetchAllMessages(payload: String, promise: Promise) {
-        bridgeHandler.fetchAllMessages(payload, promise)
-    }
-
-    @ReactMethod
-    fun deleteMessage(payload: String) {
-        bridgeHandler.deleteMessage(payload)
-    }
-
-    @ReactMethod
-    fun trackMessageClicked(payload: String) {
-        bridgeHandler.trackMessageClicked(payload)
-    }
-}
