+++ sdk/inbox/ios/MoEReactInbox.h
@@ -1,14 +1,6 @@
 
 #import <React/RCTBridgeModule.h>
-
-#ifdef RCT_NEW_ARCH_ENABLED
 #import <NativeMoEngageInboxSpec/NativeMoEngageInboxSpec.h>
-#endif
 
-#ifdef RCT_NEW_ARCH_ENABLED
 @interface MoEReactInbox : NSObject <NativeMoEngageInboxSpec>
 @end
-#else
-@interface MoEReactInbox : NSObject <RCTBridgeModule>
-@end
-#endif
