+++ sdk/inbox/ios/MoEReactInbox.mm
@@ -29,10 +29,8 @@ RCT_EXPORT_METHOD(deleteMessage:(NSString *) payload) {
     [[MoEReactNativeInboxHandler sharedInstance] deleteInboxEntry: payload];
 }
 
-#ifdef RCT_NEW_ARCH_ENABLED
 - (std::shared_ptr<facebook::react::TurboModule>)getTurboModule:(const facebook::react::ObjCTurboModule::InitParams &)params {
     return std::make_shared<facebook::react::NativeMoEngageInboxSpecJSI>(params);
 }
-#endif
 
 @end
