--- sdk/personalize/android/src/newarch/java/com/moengage/react/personalize/MoEReactPersonalize.kt
@@ -1,54 +1,0 @@
-/*
- * Copyright (c) 2026 MoEngage, Inc.
- * All rights reserved.
- * Use of source code or binaries contained within MoEngage's SDKs is permitted only to enable use of the MoEngage platform by customers of MoEngage. The Licensee may not:
- * - permit any third party to use the Software;
- * - modify or translate the Software except as otherwise permitted;
- * - reverse engineer, decompile, or disassemble the Software;
- * - copy the Software, except as expressly provided above; or
- * - remove or obscure any proprietary rights notices or labels on the Software.
- * - Licensee may not transfer the Software or any rights under this Agreement without the Licensor's prior written consent.
- * - MoEngage owns the Software and all intellectual property rights embodied therein, including copyrights and valuable trade secrets embodied in the Software. The Licensee shall not alter or remove this copyright notice.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF THE USER HAS BEEN ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-package com.moengage.react.personalize
-
-import com.facebook.react.bridge.Promise
-import com.facebook.react.bridge.ReactApplicationContext
-
-/**
- * Bridge to communicate with personalize plugin JS code in new arch.
- */
-class MoEReactPersonalize(reactContext: ReactApplicationContext) :
-    NativeMoEngagePersonalizeSpec(reactContext) {
-
-    private val bridgeHandler = MoEngagePersonalizeHandler(reactContext.applicationContext)
-
-    override fun getName() = bridgeHandler.getName()
-
-    override fun fetchExperiencesMeta(payload: String, promise: Promise) {
-        bridgeHandler.fetchExperiencesMeta(payload, promise)
-    }
-
-    override fun fetchExperiences(payload: String, promise: Promise) {
-        bridgeHandler.fetchExperiences(payload, promise)
-    }
-
-    override fun experiencesShown(payload: String) {
-        bridgeHandler.trackExperienceShown(payload)
-    }
-
-    override fun experienceClicked(payload: String) {
-        bridgeHandler.trackExperienceClicked(payload)
-    }
-
-    override fun offeringsShown(payload: String) {
-        bridgeHandler.trackOfferingShown(payload)
-    }
-
-    override fun offeringClicked(payload: String) {
-        bridgeHandler.trackOfferingClicked(payload)
-    }
-}
