+++ sdk/personalize/ios/MoEngagePersonalizeBridge.mm
@@ -37,10 +37,8 @@ RCT_EXPORT_METHOD(offeringClicked:(NSString *)payload) {
     [[MoEReactNativePersonalizeHandler sharedInstance] offeringClicked:payload];
 }
 
-#ifdef RCT_NEW_ARCH_ENABLED
 - (std::shared_ptr<facebook::react::TurboModule>)getTurboModule:(const facebook::react::ObjCTurboModule::InitParams &)params {
     return std::make_shared<facebook::react::NativeMoEngagePersonalizeSpecJSI>(params);
 }
-#endif
 
 @end
